Create the project
This is the first feature-complete and working (in my tests) version.

No unit tests because

* It's mostly glue code which is often either inconvenient to test,
  testing it is not valuable or both; on the other hand at least the CSV
  processing could use some tests
* There's generous amount of consistency checks and assertions in the
  code
* Almost everything is covered with type hints and some amount of effort
  was dedicated to use types in a way that enhanced our confidence in
  the code and prevents many classes of errors
* The potentially problematic places in the code are either visibly
  working or not, so a few manual tests are enough to discover any
  issues

Which is not to say that we shouldn't have some unit tests here if we
can think of some.

The tool is generic to a small extent and, depending on the situation,
one would have to modify it to their liking.

Several assumptions have been made in this version to the tool, it's
expected that we'll need to revisit some of them.

Linting, formatting and type checking enforced in CI.